Improves prompt processing performance with fp16 models.

3090 Ti/WSL2:
| model                          | backend    | ngl | test       |       master t/s |        PR t/s | speedup |
| ------------------------------ | ---------- | --: | ---------- | ---------------: | ------------: | ------: |
| llama 7B mostly F16            | CUDA       |  99 | pp 512     |   1661.19 ± 3.09 | 3984.28 ± 25.45 |    2.40 |