@slaren 2023-09-27
Is this actually correct? I believe compute capability 7.0 is volta, not turing.

https://github.com/ggerganov/llama.cpp/blob/7d5674dd2d045584993a46102eafa48a31388bdb/ggml-cuda.cu#L82