@bobqianic 2023-09-28
> Is this actually correct? I believe compute capability 7.0 is volta, not turing.
> 
> https://github.com/ggerganov/llama.cpp/blob/7d5674dd2d045584993a46102eafa48a31388bdb/ggml-cuda.cu#L82

**The computing capacity of Turing is 7.5, while that of Volta is 7.0. However, Volta also supports FP16.**

![image](https://github.com/ggerganov/llama.cpp/assets/129547291/ae993a78-7a1b-447a-bb4b-5157bc17b118)