@ggerganov 2023-09-28
False alarm - forgot to build with `LLAMA_CUDA_MMV_Y=2`